requirements.txt
streamlit

langchain
langchain-core
langchain-community
langchain-classic
langchain-groq
langchain-huggingface
langchain-text-splitters

sentence-transformers

faiss-cpu
transformers
torch

PyPDF2
tiktoken
pypdf

langsmith

langgraph


db-sqlite3

langgraph-checkpoint
langgraph-checkpoint-sqlite
aiosqlite


streamlit

pydantic
python-dotenv
sqlalchemy

psycopg[binary]

langgraph-checkpoint-postgres


langchain-google-genai



